--- client/src/components/AccountSettings/index.js
@@ -1,1 +1,0 @@
-export * from './Settings';
