+++ client/src/css/tailwind.css
@@ -4221,22 +4221,6 @@ table {
   margin-top: -0.25rem;
 }
 
-.-mt-2px {
-  margin-top: -2px;
-}
-
-.-mt-3px {
-  margin-top: -3px;
-}
-
-.-mt-4px {
-  margin-top: -4px;
-}
-
-.-mt-5px {
-  margin-top: -5px;
-}
-
 .-mr-1 {
   margin-right: -0.25rem;
 }
